sqlite does not support window over clause
