+++ docs/src/keywords_aggregates_rankings.md
@@ -4,7 +4,7 @@
 
 AND AS ASC BY CREATE DATABASE DELETE DESC DISTINCT DROP EXISTS FROM FULL GROUP
 HAVING IF INNER INSERT INTO IS JOIN LEFT LIKE LIMIT NOT NULL OFFSET ON OR ORDER OUTER OVER
-RIGHT SELECT SET TABLE UPDATE USING VALUES WHERE
+PARTITION RIGHT SELECT SET TABLE UPDATE USING VALUES WHERE
 
 
 ### [@aggregates](@id aggregates)
