+++ src/Adapters/sql_exports.jl
@@ -2,7 +2,7 @@
 export
     AND, AS, ASC, BY, CREATE, DATABASE, DELETE, DESC, DISTINCT, DROP, EXISTS, FROM, FULL, GROUP,
     HAVING, IF, INNER, INSERT, INTO, IS, JOIN, LEFT, LIKE, LIMIT, NOT, NULL, OFFSET, ON, OR, ORDER, OUTER, OVER,
-    RIGHT, SELECT, SET, TABLE, UPDATE, USING, VALUES, WHERE
+    PARTITION, RIGHT, SELECT, SET, TABLE, UPDATE, USING, VALUES, WHERE
 
 # export: aggregate functions
 export AVG, COUNT, SUM
@@ -13,8 +13,11 @@ export DENSE_RANK, RANK, ROW_NUMBER
 # export: PlaceHolder, Raw, Enclosed
 export PlaceHolder, Raw, Enclosed
 
-# export: Octo, Repo, Schema, from, as, window
-export Octo, Repo, Schema, from, as, window
+# export: Octo, Repo, Schema, from, as
+export Octo, Repo, Schema, from, as
+
+# export: window
+export window
 
 # export: to_sql
 export to_sql
