+++ src/Adapters/sql_imports.jl
@@ -5,7 +5,7 @@ import ...Octo
 import .Octo.AdapterBase:
     AND, AS, ASC, BY, CREATE, DATABASE, DELETE, DESC, DISTINCT, DROP, EXISTS, FROM, FULL, GROUP,
     HAVING, IF, INNER, INSERT, INTO, IS, JOIN, LEFT, LIKE, LIMIT, NOT, NULL, OFFSET, ON, OR, ORDER, OUTER, OVER,
-    RIGHT, SELECT, SET, TABLE, UPDATE, USING, VALUES, WHERE
+    PARTITION, RIGHT, SELECT, SET, TABLE, UPDATE, USING, VALUES, WHERE
 
 # import: aggregate functions
 import .Octo.AdapterBase: AVG, COUNT, SUM
@@ -15,8 +15,8 @@ import .Octo.AdapterBase: DENSE_RANK, RANK, ROW_NUMBER
 
 # import:                      ()        ?
 import .Octo.AdapterBase: Raw, Enclosed, PlaceHolder, Field
-import .Octo.AdapterBase: Database, Structured, SubQuery, _to_sql, _placeholder, _placeholders
+import .Octo.AdapterBase: Database, Structured, SubQuery, OverClause, OverClauseError, _to_sql, _placeholder, _placeholders
 
 # import: Repo, Schema, from, as
 import .Octo: Repo, Schema
-import .Octo.Queryable: from, as, window
+import .Octo.Queryable: from, as
