+++ test/adapters/jdbc/repo_test.jl
@@ -3,7 +3,7 @@ module adapters_jdbc_repo_test
 using Test # @test
 using Octo.Adapters.JDBC # Repo Schema
 
-Repo.set_log_level(Repo.LogLevelDebugSQL)
+Repo.debug_sql()
 
 Repo.connect(
     adapter = Octo.Adapters.JDBC,
