+++ test/adapters/mysql/repo_test.jl
@@ -3,7 +3,7 @@ module adapters_mysql_repo_test
 using Test # @test
 using Octo.Adapters.MySQL # Repo Schema Raw USE
 
-Repo.set_log_level(Repo.LogLevelDebugSQL)
+Repo.debug_sql()
 
 Repo.connect(
     adapter = Octo.Adapters.MySQL,
